Implement health check and move Ibc gRPC service
